:recycle: Extract provider resolution logic into helpers.rs
- Add helpers::resolve_provider() for consistent provider handling
- Remove ~15 lines of duplicated code from 7 formatter files
